+++ database.sql
@@ -26,6 +26,7 @@ CREATE TABLE IF NOT EXISTS users (
   email VARCHAR(255) UNIQUE NOT NULL,
   password_hash VARCHAR(255) NOT NULL,
   name VARCHAR(255),
+  avatar TEXT,
   is_platform_admin BOOLEAN NOT NULL DEFAULT FALSE,
   email_verified_at TIMESTAMP WITH TIME ZONE,
   created_at TIMESTAMP WITH TIME ZONE DEFAULT NOW(),